Refactor: separação dos componentes de search e notifications do sidebar e sidebarHeader
