+++ src/interface/UserInterface.ts
@@ -101,3 +101,11 @@ export interface UserPostById {
   tags: string[];
   username: string;
 }
+
+export interface UserNotifications {
+  username: string;
+  created_at: string | Date;
+  content: string;
+  post_id: UUID | string;
+  isNew: boolean;
+}
